# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Db server API description.",
        "title": "Db server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://openitstudio.ru",
            "email": "fedor@support-pc.org"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/admin/config": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "List configs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config manager",
                    "Admin"
                ],
                "summary": "List configs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/config.Config"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config manager",
                    "Admin"
                ],
                "summary": "Create config",
                "parameters": [
                    {
                        "description": "Config info",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateUserForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/user.User"
                        }
                    }
                }
            }
        },
        "/admin/config/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Config detail info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config manager",
                    "Admin"
                ],
                "summary": "Config info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Config id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Config"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config manager",
                    "Admin"
                ],
                "summary": "Update config",
                "parameters": [
                    {
                        "description": "Config info",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/config.Config"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Config id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/config.Config"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Delete config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config manager",
                    "Admin"
                ],
                "summary": "Delete config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Config id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
        "/admin/cron": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "List cron jobs",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cron",
                    "Admin"
                ],
                "summary": "List cron jobs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cron.CronJob"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Create cron job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cron",
                    "Admin"
                ],
                "summary": "Create cron job",
                "parameters": [
                    {
                        "description": "Push info",
                        "name": "cron",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cron.CronJob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cron.CronJob"
                        }
                    }
                }
            }
        },
        "/admin/cron/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Cron job detail info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cron",
                    "Admin"
                ],
                "summary": "Cron job info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Push id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cron.CronJob"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Update cron job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cron",
                    "Admin"
                ],
                "summary": "Update cron job",
                "parameters": [
                    {
                        "description": "Cron job info",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cron.CronJob"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Cron id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cron.CronJob"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "description": "Delete cron job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cron",
                    "Admin"
                ],
                "summary": "Delete cron job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cron id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
        "/config/{id}": {
            "get": {
                "description": "Get config by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config manager",
                    "Public Api"
                ],
                "summary": "Config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Config id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Auth key",
                        "name": "db-key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "cf.CloudFunction": {
            "type": "object",
            "properties": {
                "container": {
                    "description": "Container name\nexample: docker.io/library/alpine",
                    "type": "string"
                },
                "env": {
                    "description": "Container env variables",
                    "type": "string"
                },
                "id": {
                    "description": "The function UUID\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "params": {
                    "description": "Container run params\nexample: echo test",
                    "type": "string"
                },
                "project": {
                    "description": "Linked project",
                    "$ref": "#/definitions/project.Project"
                },
                "project_id": {
                    "description": "The project UUID\nexample: 6204037c-30e6-403b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "run_count": {
                    "description": "Function run count",
                    "type": "integer"
                },
                "title": {
                    "description": "Function title",
                    "type": "string"
                }
            }
        },
...
        "config.Config": {
            "type": "object",
            "properties": {
                "body": {},
                "id": {
                    "type": "string"
                },
                "project": {
                    "$ref": "#/definitions/project.Project"
                },
                "project_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "cron.CronJob": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/cf.CloudFunction"
                },
                "function_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "time_params": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "project.Project": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "origins": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "user.CreateUserForm": {
            "type": "object",
            "properties": {
                "Email": {
                    "description": "new User email",
                    "type": "string"
                },
                "Password": {
                    "description": "new User password",
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "properties": {
                "active": {
                    "description": "Is user active",
                    "type": "boolean"
                },
                "admin": {
                    "description": "Is user admin",
                    "type": "boolean"
                },
                "created_at": {
                    "description": "Created at date time",
                    "type": "string"
                },
                "devices": {
                    "description": "user devices",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.UserDevice"
                    }
                },
                "email": {
                    "description": "User email",
                    "type": "string"
                },
                "id": {
                    "description": "The user UUID\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "last_login": {
                    "description": "Last login date time",
                    "type": "string"
                },
                "token": {
                    "description": "Auth token",
                    "type": "string"
                }
            }
        },
        "user.UserDevice": {
            "type": "object",
            "properties": {
                "device": {
                    "description": "Device type can be android | ios | macos | windows | linux | web\nexample: 'linux'",
                    "type": "string"
                },
                "device_token": {
                    "description": "UUID Device Token\n9f80fcc8-0102-4795-94c0-4190c168ffc2",
                    "type": "string"
                },
                "id": {
                    "description": "The UUID of a device\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                },
                "user_id": {
                    "description": "The UUID of owned user\nexample: 6204037c-30e6-408b-8aaa-dd8219860b4b",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}